+++ terratex_reallife/SYSTEM/vehsys/start_veh.lua
@@ -229,8 +229,24 @@ addEventHandler("respawnVehicleByAdmin", getRootElement(), respawntheVehiclebyad
 function deletetheVehiclebyadmin_func(vehicle, grund)
     if (isAdminLevel(source, 1)) then
         local id = vioGetElementData(vehicle, "dbid")
+        local besitzer = vioGetElementData(vehicle, "besitzer");
         triggerClientEvent(source, "showErrorText", source, source, "Das Fahrzeug wurde erfolgreich geloescht!")
-        save_offline_message(vioGetElementData(vehicle, "besitzer"), getPlayerName(source), string.format("Dein Fahrzeug in Slot %s wurde gelöscht, weil: %s", vioGetElementData(vehicle, "slotid"), grund))
+
+        local slot = vioGetElementData(vehicle,"slotid")
+
+        privCars[vehicle] = false
+        for theKey, theTable in ipairs(privVeh) do
+            if (theTable[3] == vehicle) then
+                table.remove(privVeh, theKey)
+            end
+        end
+
+        if (getPlayerFromName(besitzer)) then
+            vioSetElementData(besitzer, "slot" .. SlotID, -1)
+            outputChatBox(string.format("Dein Fahrzeug in Slot %s wurde gelöscht, weil: %s", vioGetElementData(vehicle, "slotid"), grund), besitzer)
+        else
+            save_offline_message(vioGetElementData(vehicle, "besitzer"), getPlayerName(source), string.format("Dein Fahrzeug in Slot %s wurde gelöscht, weil: %s", vioGetElementData(vehicle, "slotid"), grund))
+        end
 
         MySql.helper.delete("user_vehicles", {ID = id});
         elementData[vehicle] = nil